refactor: reorganize profile and score tabs into separate screens and update imports
